+++ monks-little-details.js
@@ -929,12 +929,18 @@ Hooks.on("getActorDirectoryEntryContext", (html, entries) => {
             icon: '<i class="fas fa-random"></i>',
             condition: li => {
                 const actor = game.actors.get(li.data("documentId"));
-                const canPolymorph = game.user.isGM || (actor.isOwner && game.settings.get("dnd5e", "allowPolymorphing"));
+                const canPolymorph = (game.user.isGM || (actor.testUserPermission(game.user, "OWNER") && game.user.can("TOKEN_CREATE"))) && game.settings.get("dnd5e", "allowPolymorphing");
                 return canPolymorph;
             },
             callback: async (li) => {
+                let docId = li.data("documentId");
+                let from = game.actors.get(docId);
+
+                if (!from) return;
+
                 let data = {
-                    id: li.data("documentId")
+                    type: 'Actor',
+                    uuid: from.uuid
                 }
 
                 let actors = canvas.tokens.controlled.map(t => t.actor);
@@ -1082,4 +1088,51 @@ Hooks.on("renderFilePicker", (app, html, data) => {
         )
         .after(list);
     $('input[name="target"]', html).parent().css({position: "relative"});
+});
+
+Hooks.on('renderCompendiumDirectory', (app, html, data, options) => {
+    if (setting("compendium-shortcuts")) {
+        let shortcut = $('<div>').addClass('action-buttons flexrow').append(`
+        <nav class="tabs compendium-shortcut-links">
+            <a class="item" data-tab="Actor" data-tooltip="DOCUMENT.Actors" alt="DOCUMENT.Actors">
+                <i class="fas fa-user"></i>
+            </a>
+            <a class="item" data-tab="Adventure" data-tooltip="DOCUMENT.Adventures" alt="DOCUMENT.Adventures">
+                <i class="fas fa-map-pin"></i>
+            </a>
+            <a class="item" data-tab="Cards" data-tooltip="DOCUMENT.CardsPlural" alt="DOCUMENT.CardsPlural">
+                <i class="fa-solid fa-cards"></i>
+            </a>
+            <a class="item" data-tab="Item" data-tooltip="DOCUMENT.Items" alt="DOCUMENT.Items">
+                <i class="fas fa-suitcase"></i>
+            </a>
+            <a class="item" data-tab="JournalEntry" data-tooltip="DOCUMENT.JournalEntries" alt="DOCUMENT.JournalEntries">
+                <i class="fas fa-book-open"></i>
+            </a>
+            <a class="item" data-tab="Macro" data-tooltip="DOCUMENT.Macros" alt="DOCUMENT.Macros">
+                <i class="fas fa-code"></i>
+            </a>
+            <a class="item" data-tab="Playlist" data-tooltip="DOCUMENT.Playlists" alt="DOCUMENT.Playlists">
+                <i class="fas fa-music"></i>
+            </a>
+            <a class="item" data-tab="RollTable" data-tooltip="DOCUMENT.RollTables" alt="DOCUMENT.RollTables">
+                <i class="fas fa-th-list"></i>
+            </a>
+            <a class="item" data-tab="Scene" data-tooltip="DOCUMENT.Scenes" alt="DOCUMENT.Scenes">
+                <i class="fas fa-map"></i>
+            </a>
+        </nav>`);
+        $('.item', shortcut).on("click", (evt) => {
+            let id = evt.currentTarget.dataset.tab;
+            let title = $(`h3:contains(${id}),li.compendium-type[data-type="${id}"]`, app.element);
+            if (title.length) {
+                title[0].scrollIntoView({ behavior: "smooth", block: "start" });
+            }
+        }).each((idx, elem) => {
+            let id = elem.dataset.tab;
+            let title = $(`h3:contains(${id}),li.compendium-type[data-type="${id}"]`, app.element);
+            $(elem).toggleClass("disabled", !title.length);
+        });
+        $('.directory-header', html).append(shortcut);
+    }
 });